Update item upgrade, drop item, portal ui
Hqua quên up nên hôm nay viết cái này chịu

+++ Assets/Scripts/UI/MonsterDamageUI.cs
@@ -10,7 +10,7 @@ public class MonsterDamageUI : MonoBehaviour
 
 
 
-    public void DamageUI( int damage ) {
+    public void DamageUI( float damage ) {
         
         damageText.text = "-" + damage.ToString();
         TextMeshProUGUI dmgText = Instantiate( damageText, transform.position, Quaternion.identity, transform);
